+++ NEWS.md
@@ -33,6 +33,7 @@ None
 26. Fix #75 - Add function `util_beta_param_estimate()`
 27. Fix #76 - Add function `util_binomial_param_estimate()`
 28. Fix #77 - Add function `util_exponential_param_estimate()`
+29. Fix #78 - Add function `util_gamma_param_estimate()`
 
 ## Fixes and Minor Improvements
 1. Fix #30 - Move `crayon`, `rstudioapi`, and `cli` from Suggests to Imports due to `pillar`